requirements.txt
fastapi
uvicorn
pywebview